requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==4.2.21
exceptiongroup==1.3.0
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.78.1
openai-agents==0.0.14
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
